fix: use shadow root (CT-000)
