--- packages/react-chat/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
-    <title>Voiceflow Widget</title>
-    <style>
-      body,
-      html,
-      #vfchat {
-        margin: 0;
-        height: 100%;
-        width: 100%;
-        -moz-osx-font-smoothing: grayscale;
-        -webkit-font-smoothing: antialiased;
-        touch-action: manipulation;
-      }
-    </style>
-    <script type="module" src="/src/browser/index.tsx"></script>
-  </head>
-  <body></body>
-</html>
