--- packages/react-chat/index.ts
@@ -1,9 +1,0 @@
-export * from './src/common';
-export * from './src/components';
-export * from './src/contexts';
-export * from './src/hooks';
-export * from './src/runtime';
-export * from './src/styles';
-export * from './src/types/turn';
-export * from './src/utils/functional';
-export * from './src/views';
