+++ packages/react-chat/package.json
@@ -13,6 +13,7 @@
     "@voiceflow/sdk-runtime": "1.7.0",
     "@voiceflow/slate-serializer": "1.5.5",
     "@voiceflow/voiceflow-types": "3.26.9",
+    "bowser": "^2.11.0",
     "chroma-js": "2.4.2",
     "clsx": "1.2.1",
     "csstype": "3.1.0",
@@ -68,6 +69,7 @@
     "vite": "3.2.2",
     "vite-plugin-dts": "^1.6.6",
     "vite-plugin-fonts": "^0.6.0",
+    "vite-plugin-html": "^3.2.0",
     "vite-plugin-svgr": "^2.2.1",
     "vite-tsconfig-paths": "^3.5.0",
     "vitest": "0.23.1"
@@ -93,7 +95,6 @@
     "build": "yarn clean && yarn build:package && yarn build:bundle",
     "build:bundle": "NODE_ENV=production vite build",
     "build:docs": "NODE_OPTIONS=--openssl-legacy-provider STORYBOOK_BASE_HREF=react-chat build-storybook -o docs",
-    "build:package": "NODE_ENV=production vite --config vite.package.config.ts build && tsc-alias -p tsconfig.build.json",
     "clean": "rimraf build dist",
     "dev": "NODE_ENV=development vite",
     "eslint-output": "eslint-output",
