+++ packages/react-chat/src/common/index.ts
@@ -1,5 +1,3 @@
 export * from './hooks';
-export * as Listeners from './listeners';
-export * as PostMessage from './postMessage';
 export * from './types';
 export * from './utils';
