+++ packages/react-chat/src/components/SystemResponse/index.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 
 import type { RuntimeAction } from '@/common';
 import Button from '@/components/Button';
-import { RuntimeAPIContext } from '@/contexts';
+import { RuntimeContext } from '@/contexts';
 import { useAutoScroll } from '@/hooks';
 
 import Feedback, { FeedbackProps } from '../Feedback';
@@ -60,7 +60,7 @@ export interface SystemResponseProps {
 }
 
 const SystemResponse: React.FC<SystemResponseProps> = ({ feedback, avatar, timestamp, messages, actions = [], isLast, Message = SystemMessage }) => {
-  const runtime = useContext(RuntimeAPIContext);
+  const runtime = useContext(RuntimeContext);
 
   const { showIndicator, visibleMessages, complete } = useAnimatedMessages({
     messages,
@@ -86,8 +86,8 @@ const SystemResponse: React.FC<SystemResponseProps> = ({ feedback, avatar, times
 
       {isLast && complete && !!actions.length && (
         <Actions>
-          {actions.map(({ name, request }, index) => (
-            <Button variant="secondary" onClick={() => runtime?.send(name, request)} key={index}>
+          {actions.map(({ request, name }, index) => (
+            <Button variant="secondary" onClick={() => runtime?.interact(request, name)} key={index}>
               {name}
             </Button>
           ))}
