+++ packages/react-chat/src/contexts/index.ts
@@ -1,2 +1,2 @@
 export * from './AutoScrollContext';
-export * from './RuntimeAPIContext';
+export * from './RuntimeContext';
