+++ packages/react-chat/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useAutoScroll';
 export * from './useDidUpdateEffect';
-export * from './useRuntime';
 export * from './useStateRef';
