+++ packages/react-chat/src/utils/session.ts
