+++ packages/react-chat/src/views/ChatWindow/index.tsx
@@ -1,47 +1,25 @@
 import '../../styles.css';
 
-import React, { useCallback } from 'react';
+import React, { useCallback, useContext } from 'react';
 import * as R from 'remeda';
 import { match } from 'ts-pattern';
 
-import { Assistant, ChatConfig, Listeners, PostMessage, SessionOptions, SessionStatus, useTheme } from '@/common';
+import { SessionStatus, useTheme } from '@/common';
 import { Chat, SystemResponse, UserResponse } from '@/components';
-import { RuntimeAPIProvider } from '@/contexts';
-import { FeedbackName, useRuntime } from '@/hooks';
+import { RuntimeContext } from '@/contexts/RuntimeContext';
+import type { FeedbackName } from '@/contexts/RuntimeContext/useRuntimeAPI';
 import { TurnType, UserTurnProps } from '@/types';
-import { useResolveAssistantStyleSheet } from '@/utils/stylesheet';
 
 import { ChatWindowContainer } from './styled';
 
-const ChatWindow: React.FC<ChatConfig & { assistant: Assistant; session: SessionOptions }> = ({
-  assistant,
-  versionID,
-  verify,
-  user,
-  url,
-  session,
-  launch,
-}) => {
+const ChatWindow: React.FC = () => {
+  const runtime = useContext(RuntimeContext);
+  const { assistant } = runtime;
+
   // emitters
-  const close = useCallback(() => Listeners.sendMessage({ type: PostMessage.Type.CLOSE }), []);
-  const saveSession = useCallback((session: SessionOptions) => Listeners.sendMessage({ type: PostMessage.Type.SAVE_SESSION, payload: session }), []);
-
-  const runtime = useRuntime({ versionID, verify, url, user, session, saveSession });
-
-  // listeners
-  Listeners.useListenMessage(PostMessage.Type.OPEN, async (): Promise<void> => {
-    if (runtime.isStatus(SessionStatus.IDLE)) {
-      await handleStart();
-    }
-  });
-
-  const handleStart = async (): Promise<void> => {
-    await runtime.launch(launch);
-  };
-
   const closeAndEnd = useCallback((): void => {
     runtime.setStatus(SessionStatus.ENDED);
-    close();
+    runtime.close();
   }, []);
 
   const theme = useTheme(assistant);
@@ -54,52 +32,47 @@ const ChatWindow: React.FC<ChatConfig & { assistant: Assistant; session: Session
     [runtime.session.turns]
   );
 
-  const isStyleSheetResolved = useResolveAssistantStyleSheet(assistant);
-  if (!isStyleSheetResolved) return null;
-
   return (
-    <RuntimeAPIProvider {...runtime}>
-      <ChatWindowContainer className={theme}>
-        <Chat
-          title={assistant.title}
-          description={assistant.description}
-          image={assistant.image}
-          avatar={assistant.avatar}
-          withWatermark={assistant.watermark}
-          startTime={runtime.session.startTime}
-          hasEnded={runtime.isStatus(SessionStatus.ENDED)}
-          isLoading={!runtime.session.turns.length}
-          onStart={handleStart}
-          onEnd={closeAndEnd}
-          onSend={runtime.reply}
-          onMinimize={close}
-        >
-          {runtime.session.turns.map((turn, turnIndex) =>
-            match(turn)
-              .with({ type: TurnType.USER }, ({ id, ...props }) => <UserResponse {...R.omit(props, ['type'])} key={id} />)
-              .with({ type: TurnType.SYSTEM }, ({ id, ...props }) => (
-                <SystemResponse
-                  key={id}
-                  {...R.omit(props, ['type'])}
-                  feedback={
-                    assistant.feedback
-                      ? {
-                          onClick: (feedback: FeedbackName) => {
-                            runtime.feedback(feedback, props.messages, getPreviousUserTurn(turnIndex));
-                          },
-                        }
-                      : undefined
-                  }
-                  avatar={assistant.avatar}
-                  isLast={turnIndex === runtime.session.turns.length - 1}
-                />
-              ))
-              .exhaustive()
-          )}
-          {runtime.indicator && <SystemResponse.Indicator avatar={assistant.avatar} />}
-        </Chat>
-      </ChatWindowContainer>
-    </RuntimeAPIProvider>
+    <ChatWindowContainer className={theme}>
+      <Chat
+        title={assistant.title}
+        description={assistant.description}
+        image={assistant.image}
+        avatar={assistant.avatar}
+        withWatermark={assistant.watermark}
+        startTime={runtime.session.startTime}
+        hasEnded={runtime.isStatus(SessionStatus.ENDED)}
+        isLoading={!runtime.session.turns.length}
+        onStart={runtime.launch}
+        onEnd={closeAndEnd}
+        onSend={runtime.reply}
+        onMinimize={runtime.close}
+      >
+        {runtime.session.turns.map((turn, turnIndex) =>
+          match(turn)
+            .with({ type: TurnType.USER }, ({ id, ...props }) => <UserResponse {...R.omit(props, ['type'])} key={id} />)
+            .with({ type: TurnType.SYSTEM }, ({ id, ...props }) => (
+              <SystemResponse
+                key={id}
+                {...R.omit(props, ['type'])}
+                feedback={
+                  assistant.feedback
+                    ? {
+                        onClick: (feedback: FeedbackName) => {
+                          runtime.feedback(feedback, props.messages, getPreviousUserTurn(turnIndex));
+                        },
+                      }
+                    : undefined
+                }
+                avatar={assistant.avatar}
+                isLast={turnIndex === runtime.session.turns.length - 1}
+              />
+            ))
+            .exhaustive()
+        )}
+        {runtime.indicator && <SystemResponse.Indicator avatar={assistant.avatar} />}
+      </Chat>
+    </ChatWindowContainer>
   );
 };
