--- packages/react-chat/vite.package.config.ts
@@ -1,31 +1,0 @@
-import react from '@vitejs/plugin-react';
-import path from 'path';
-import { defineConfig } from 'vite';
-import dts from 'vite-plugin-dts';
-
-import { createPlugins } from './vite.config';
-
-export default defineConfig({
-  define: {
-    'process.env': '({})',
-  },
-  build: {
-    outDir: path.resolve(__dirname, 'build'),
-    lib: {
-      entry: path.resolve(__dirname, 'index.ts'),
-      name: 'voiceflow-chat',
-      fileName: (format) => `index.${format}.js`,
-      formats: ['es', 'cjs'],
-    },
-    rollupOptions: {
-      external: ['react', 'react-dom'],
-      output: {
-        globals: {
-          react: 'React',
-          'react-dom': 'ReactDOM',
-        },
-      },
-    },
-  },
-  plugins: [react(), dts(), ...createPlugins()],
-});
