--- packages/widget/.env.development
@@ -1,4 +1,0 @@
-# THIS IS A PUBLIC FILE, THIS WILL BE COMMITTED TO GITHUB
-
-VITE_WIDGET_URL='http://localhost:3006'
-VITE_VF_RUNTIME_URL='https://general-runtime.development.voiceflow.com'
