--- packages/widget/.env.production
@@ -1,4 +1,0 @@
-# THIS IS A PUBLIC FILE, THIS WILL BE COMMITTED TO GITHUB
-
-# dotenv-expand doesn't allow referencing of global variables, CIRCLE_SHA1 gets replaced in vite.config.ts
-VITE_WIDGET_URL='https://cdn.voiceflow.com/react-chat/CIRCLE_SHA1/index.html'
