--- packages/widget/.eslintignore
@@ -1,7 +1,0 @@
-coverage
-.nyc_output
-/build
-/dist
-/docs
-/node_modules
-!.storybook
