--- packages/widget/.eslintoutputrc
@@ -1,13 +1,0 @@
-{
-  "formats": [
-    {
-      "name": "stylish",
-      "output": "console"
-    },
-    {
-      "name": "json",
-      "output": "file",
-      "path": "sonar/report.json"
-    }
-  ]
-}
