--- packages/widget/.gitignore
@@ -1,2 +1,0 @@
-.env.local
-.env.*.local
