--- packages/widget/examples/index.html
@@ -1,29 +1,0 @@
-<!-- THIS IS AN EXAMPLE PAGE AND SHOULD NOT BE USED OUTSIDE OF DEVELOPMENT -->
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Example Page</title>
-  <style>
-    body {
-      background-color: #f9f9f9;
-    }
-  </style>
-</head>
-
-<body>
-  <script>
-    function _vf_load() {
-      window.voiceflow.chat.load({
-        url: '<%= VITE_VF_RUNTIME_URL %>',
-        verify: { projectID: '<%= VITE_VF_PROJECT_ID %>' },
-        user: {
-          name: 'Development User',
-        },
-      });
-    }
-  </script>
-  <script type="module" src="../index.tsx" onload="_vf_load()"></script>
-</body>
-
-</html>
