--- packages/widget/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "@voiceflow/widget",
-  "description": "voiceflow iframe widget",
-  "version": "1.7.17",
-  "author": "Tyler Han",
-  "bugs": {
-    "url": "https://github.com/voiceflow/react-chat/issues"
-  },
-  "dependencies": {
-    "@voiceflow/react-chat": "workspace:*",
-    "@voiceflow/sdk-runtime": "1.7.0",
-    "bowser": "^2.11.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "17.0.3",
-    "@emotion/core": "10.1.1",
-    "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.6",
-    "@vitejs/plugin-react": "^2.0.1",
-    "@vitest/coverage-c8": "^0.23.1",
-    "@voiceflow/base-types": "^2.96.3",
-    "@voiceflow/eslint-config": "6.1.0",
-    "@voiceflow/prettier-config": "1.2.1",
-    "@voiceflow/sdk-runtime": "1.7.0",
-    "@voiceflow/tsconfig": "1.4.8",
-    "babel-loader": "^8.2.5",
-    "depcheck": "1.4.3",
-    "eslint": "7.32.0",
-    "eslint-output": "^3.0.1",
-    "eslint-plugin-mdx": "^2.0.2",
-    "fixpack": "^4.0.0",
-    "husky": "^8.0.0",
-    "lint-staged": "13.0.3",
-    "prettier": "2.7.1",
-    "type-fest": "^4.6.0",
-    "vite": "3.2.2",
-    "vite-plugin-html": "3.2.0",
-    "vite-tsconfig-paths": "^3.5.0"
-  },
-  "files": [
-    "dist/"
-  ],
-  "homepage": "https://github.com/voiceflow/react-chat#readme",
-  "keywords": [
-    "chat widget",
-    "iframe",
-    "voiceflow"
-  ],
-  "license": "MIT",
-  "main": "index.js",
-  "prettier": "@voiceflow/prettier-config",
-  "scripts": {
-    "build": "NODE_ENV=production vite build",
-    "dev": "NODE_ENV=development vite",
-    "lint": "eslint '**/*.{js,ts,tsx,mdx}'",
-    "lint:output": "yarn run eslint-output --quiet \"**/*.{js,ts,tsx,mdx}\"",
-    "lint:quiet": "yarn lint --quiet",
-    "lint:report": "yarn lint:output",
-    "test:dependencies": "depcheck"
-  }
-}
