--- packages/widget/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "extends": "@voiceflow/tsconfig/web.json",
-  "ts-node": {
-    "compilerOptions": {
-      "module": "commonjs"
-    }
-  },
-  "compilerOptions": {
-    "skipLibCheck": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "lib": [
-      "dom",
-      "es2019"
-    ],
-    "jsx": "react-jsx",
-    "baseUrl": ".",
-    "outDir": "build",
-    "types": [
-      "node"
-    ]
-  },
-  "compileOnSave": false,
-  "exclude": [
-    "node_modules",
-    "build",
-    "dist",
-    "docs"
-  ]
-}
