--- packages/widget/typings/env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="vite-plugin-svgr/client" />
-/// <reference types="vite/client" />
